ypsiloid: move YamlAst into Yaml; name YAML-spec types
Moves the `YamlAst` opaque type (and its `extends Format` companion object,
which carries `Position`, `Issue`, `Byte`, factories, extractors, and the
deep-equality helpers) inside `object Yaml` as `Yaml.Ast`. Adds transparent
type aliases corresponding to YAML 1.2 spec primitives so the `Ast` union
reads as documentation:

  type YamlString    = String
  type YamlInteger   = Long | Bcd
  type YamlDecimal   = Double | Bcd
  type YamlBoolean   = Boolean
  type YamlNull      = Null
  type YamlSequence  = IArray[Any]
  type YamlMapping   = IArray[Any]

`Bcd` deliberately appears in both `YamlInteger` and `YamlDecimal` because
the same JVM representation carries either depending on the source token;
the union collapses at type level, and the aliases stay documentary.

Aligns with PR #1065 (Jacinta) and PR #1063 (Breviloquence). The module
collapse from those PRs doesn't apply: ypsiloid was already a single
`core` module.
