+++ lib/ypsiloid/src/test/ypsiloid_conformance.scala
@@ -308,21 +308,21 @@ object Conformance:
 
       case Outcome.Passed => ()
 
-  private def yamlAstToJson(yaml: YamlAst): Json = yaml match
-    case YamlAst.Null         => Json.ast(Json.Ast(null))
-    case YamlAst.Bool(b)      => Json.ast(Json.Ast(b))
-    case YamlAst.Integer(n)   => Json.ast(Json.Ast(n))
-    case YamlAst.Decimal(d)   => Json.ast(Json.Ast(d))
-    case YamlAst.Str(s)       => Json.ast(Json.Ast(s.s))
+  private def yamlAstToJson(yaml: Yaml.Ast): Json = yaml match
+    case Yaml.Ast.Null         => Json.ast(Json.Ast(null))
+    case Yaml.Ast.Bool(b)      => Json.ast(Json.Ast(b))
+    case Yaml.Ast.Integer(n)   => Json.ast(Json.Ast(n))
+    case Yaml.Ast.Decimal(d)   => Json.ast(Json.Ast(d))
+    case Yaml.Ast.Str(s)       => Json.ast(Json.Ast(s.s))
 
-    case YamlAst.Sequence(items) =>
+    case Yaml.Ast.Sequence(items) =>
       val converted: IArray[Any] =
         IArray.from(items.map(item => Json.unseal(yamlAstToJson(item)).asInstanceOf[Any]))
       Json.ast(Json.Ast.arr(converted))
 
-    case YamlAst.Mapping(entries) =>
+    case Yaml.Ast.Mapping(entries) =>
       val pairs = entries.collect:
-        case (YamlAst.Str(s), v) => (s.s, Json.unseal(yamlAstToJson(v)))
+        case (Yaml.Ast.Str(s), v) => (s.s, Json.unseal(yamlAstToJson(v)))
 
       val keys: IArray[String] = IArray.from(pairs.map(_._1))
       val values: IArray[Any] = IArray.from(pairs.map(_._2.asInstanceOf[Any]))
